requirements.txt
requests
httpx
python-dotenv
openpyxl
aiofiles
